pyproject.toml
[project]
name = "karierai"
version = "0.3.0"
description = "KarierAI - multi-agent AI career assistant platform"
authors = [
    {name = "OpenAI Refactor", email = "noreply@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.109.0,<0.110.0)",
    "uvicorn (>=0.27.0,<0.28.0)",
    "streamlit (>=1.52.2,<2.0.0)",
    "requests (>=2.32.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.7.0,<3.0.0)",
    "langchain (>=1.2.0,<2.0.0)",
    "langchain-openai (>=1.1.5,<2.0.0)",
    "langchain-qdrant (>=1.1.0,<2.0.0)",
    "qdrant-client (>=1.9.0,<2.0.0)",
    "langfuse (>=3.11.0,<4.0.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)",
    "pymupdf (>=1.24.0,<2.0.0)",
    "pytesseract (>=0.3.10,<1.0.0)",
    "pillow (>=11.0.0,<12.0.0)",
]

[tool.poetry]
packages = [{include = "karierai", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0,<9.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
